moving promise overlap avoiding helper to a separate file
